requirements.txt
streamlit==1.32.0
opencv-python==4.9.0.80
torch==2.2.0
torchvision==0.17.0
numpy==1.26.4
Pillow==10.2.0
pandas==2.2.1 